+++ web/src/components/draft/create/index.ts
@@ -1,5 +1,9 @@
-export { DraftActionFooter } from "./draft-action-footer";
-export { DraftOptionCard } from "./draft-option-card";
-export { DraftRoomSettingsForm } from "./draft-room-settings-form";
-export { DraftSettingsSummary } from "./draft-settings-summary";
-export { DraftStepper, type DraftCreateStep } from "./draft-stepper";
+export { DraftActionFooter } from "./common/draft-action-footer";
+export { DraftStepper, type DraftCreateStep } from "./common/draft-stepper";
+export { DraftOptionCard } from "./settings/draft-option-card";
+export { DraftRoomSettingsForm } from "./settings/draft-room-settings-form";
+export { DraftSettingsSummary } from "./settings/draft-settings-summary";
+export { DraftStreamerBasicSettingsSection } from "./streamers/draft-streamer-basic-settings-section";
+export { DraftStreamerBoardSection } from "./streamers/draft-streamer-board-section";
+export { DraftStreamerPartyShareSection } from "./streamers/draft-streamer-party-share-section";
+export { DraftStreamerSearchSection } from "./streamers/draft-streamer-search-section";
